requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
flask-migrate==4.0.5
flask-mail==0.9.1
apscheduler==3.10.1